[clang-tidy] Rename google-readability-casting to modernize-avoid-c-style-cast The old name is kept as an alias for backward compatibility.
Fixes #170978
