+++ clang-tools-extra/docs/clang-tidy/checks/google/readability-casting.rst
@@ -3,12 +3,4 @@
 google-readability-casting
 ==========================
 
-Finds usages of C-style casts.
-
-https://google.github.io/styleguide/cppguide.html#Casting
-
-Corresponding cpplint.py check name: `readability/casting`.
-
-This check is similar to ``-Wold-style-cast``, but it suggests automated fixes
-in some cases. The reported locations should not be different from the
-ones generated by ``-Wold-style-cast``.
+The `google-readability-casting` check is an alias, please see `modernize-avoid-c-style-cast <../modernize/avoid-c-style-cast.html>`_ for more information.
